requirements.txt
annotated-types==0.7.0
anyio==4.8.0
azure-ai-inference==1.0.0b9
azure-core==1.32.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
isodate==0.7.2
jiter==0.8.2
openai==1.63.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
six==1.17.0
sniffio==1.3.1
tiktoken==0.9.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
